Revert default `objSortStrategy` to `.averageScore`
In 6cde912 `objSortStrategy` was introduced, allowing users to choose whether to sort by average score or minimal score. Prior to that commit, Ifrit would always sort objects by their average score.

But the commit also changed the default `objSortStrategy` to `.minimalScore`, breaking backwards compatibility and breaking one of the automated tests.

This commit sets the default strategy to `.averageScore`, as it used to be, and fixes the automated test.

Fixes #11