Change default `objSortStrategy` to `.minimalScore`
**Backwards incompatible, breaking change**

In 6cde912 `objSortStrategy` was introduced, allowing users to choose whether to sort by average score or minimal score. Prior to that commit, Ifrit would always sort objects by their average score.

But the commit also changed the default `objSortStrategy` to `.minimalScore`, breaking backwards compatibility and breaking one of the automated tests.

In #12, we reverted the default to `.averageScore`, but `.minimalScore` is more intuitive.

This commit un-reverts the default back to `.minimalScore`, and updates the test to match the new expectation.